RED: Adapt TestBidsPreprocessing to use fetch_bids_dataset()
Add fetch_bids_dataset() stub that creates BIDS directory structure
without NIfTI files, and update tests to consume it:
- Add fetch_bids_dataset() to fetch.py and data/__init__.py
- Add bids_dataset and small_ref_path_for_bids fixtures to conftest.py
- Simplify TestBidsPreprocessing._setup_bids to use bids_dataset fixture
- Remove manual BIDS directory construction from test_preprocess.py

Tests fail as expected (RED): BIDSLayout finds no subjects because
NIfTI files are not yet copied into the BIDS structure.
