+++ src/anat2vessels/data/__init__.py
@@ -1,4 +1,5 @@
 from anat2vessels.data.fetch import (  # noqa: F401
+    fetch_bids_dataset,
     fetch_ref_img,
     fetch_test_data,
     ensure_model_installed,
